Roles-Permission-dashboardAddLink-paymentMethChanges

+++ app/Http/Controllers/AssignMakeModelController.php
@@ -6,6 +6,8 @@ use App\Models\Year;
 use App\Models\VModel;
 use Illuminate\Http\Request;
 use App\Models\VehicleVeriant;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class AssignMakeModelController extends Controller
@@ -15,8 +17,15 @@ class AssignMakeModelController extends Controller
      */
     public function index()
     {
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        $assigned = $error = null;
+        if($role->hasPermissionTo('view-assign-vehicle-variant')) {
         $assigned = VehicleVeriant::with(['year','model.make'])->get();
-        return view('admin.seeAssigned',['assign'=> $assigned]);
+        } else {
+            $error = "Permission Denied";
+        }
+        return view('admin.seeAssigned',['assign'=> $assigned , 'error' => $error]);
     }
 
     /**
@@ -35,24 +44,30 @@ class AssignMakeModelController extends Controller
      */
     public function store(Request $request)
     {
-        $validation = Validator::make($request->all(),[
-            'year'=>'required',
-            'model'=>'required'
-        ]);
-        if($validation->fails()){
-            return response()->json(['status'=> 500 , 'errors' => $validation->errors()]);
-        } else {
-            foreach ($request['year'] as $yearId) {
-                foreach ($request['model'] as $modelId) {
-                    $insert = VehicleVeriant::updateOrCreate([
-                        'year_id' => $yearId,
-                        'model_id' => $modelId,
-                    ]);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('create-assign-vehicle-variant')) {
+            $validation = Validator::make($request->all(),[
+                'year'=>'required',
+                'model'=>'required'
+            ]);
+            if($validation->fails()){
+                return response()->json(['status'=> 500 , 'errors' => $validation->errors()]);
+            } else {
+                foreach ($request['year'] as $yearId) {
+                    foreach ($request['model'] as $modelId) {
+                        $insert = VehicleVeriant::updateOrCreate([
+                            'year_id' => $yearId,
+                            'model_id' => $modelId,
+                        ]);
+                    }
+                }
+                if($insert) {
+                    return response()->json(['status' => 200 , 'message' => 'Assigned Successfully']);
                 }
             }
-            if($insert) {
-                return response()->json(['status' => 200 , 'message' => 'Assigned Successfully']);
-            }
+        } else {
+            return response()->json(['status' => 500 , 'errors' => "Permission Denied"]);
         }
        
     }
@@ -86,11 +101,17 @@ class AssignMakeModelController extends Controller
      */
     public function destroy(string $id)
     {
-        $delete = VehicleVeriant::where('id' , $id)->delete();
-        if($delete) {
-            return response()->json(['status' => 200 , 'message' => 'Deleted Successfully']);
-        } else {
-            return response()->json(['status' => 500 , 'message' => "something went wrong"]);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('delete-assign-vehicle-variant')) {
+            $delete = VehicleVeriant::where('id' , $id)->delete();
+            if($delete) {
+                return response()->json(['status' => 200 , 'message' => 'Deleted Successfully']);
+            } else {
+                return response()->json(['status' => 500 , 'message' => "something went wrong"]);
+            }
+        } else{
+            return response()->json(['status' => 500 , 'message' => "Permission Denied"]);
         }
     }
 }
