+++ app/Http/Controllers/MakeController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Make;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class MakeController extends Controller
@@ -13,9 +15,15 @@ class MakeController extends Controller
      */
     public function index()
     {
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        $make = $error = null;
+        if($role->hasPermissionTo('view-make')) {
         $make = Make::all();
-        // dd($make);
-        return view('admin.manageMake',['make' => $make]);
+        } else {
+            $error = "You don't have access to this page";
+        }
+        return view('admin.manageMake',['make' => $make,'error' => $error]);
     }
 
     /**
@@ -31,17 +39,23 @@ class MakeController extends Controller
      */
     public function store(Request $request)
     {
-        $validation = Validator::make($request->all(),[
-            'make' => 'required',
-            'slug' => 'required|unique:make,slug,'
-        ]);
-        if($validation->passes()){
-            $insert = Make::create(['name'=> $request->make , 'slug' => $request->slug ]);
-            if($insert) {
-                return response()->json(['status' => 200, 'message' => 'Make Added Successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('create-make')){
+            $validation = Validator::make($request->all(),[
+                'make' => 'required',
+                'slug' => 'required|unique:make,slug,'
+            ]);
+            if($validation->passes()){
+                $insert = Make::create(['name'=> $request->make , 'slug' => $request->slug ]);
+                if($insert) {
+                    return response()->json(['status' => 200, 'message' => 'Make Added Successfully']);
+                }
+            } else {
+                return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
             }
         } else {
-            return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
+            return response()->json(['status' => 500, 'errors' => "Permission Denied"]);
         }
     }
 
@@ -66,17 +80,23 @@ class MakeController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $validation = Validator::make($request->all(),[
-            'make' => 'required',
-            'slug' => 'required|unique:make,slug,'.$id
-        ]);
-        if($validation->passes()) {
-            $update = Make::where('id', $id)->update(['name' => $request->make , 'slug' => $request->slug ]);
-            if($update) {
-                return response()->json(['status' => 200, 'message' => 'Make Updated Successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('update-make')){
+            $validation = Validator::make($request->all(),[
+                'make' => 'required',
+                'slug' => 'required|unique:make,slug,'.$id
+            ]);
+            if($validation->passes()) {
+                $update = Make::where('id', $id)->update(['name' => $request->make , 'slug' => $request->slug ]);
+                if($update) {
+                    return response()->json(['status' => 200, 'message' => 'Make Updated Successfully']);
+                }
+            } else {
+                return response()->json(['status' => 500 , 'errors' => $validation->errors() ]);
             }
         } else {
-            return response()->json(['status' => 500 , 'errors' => $validation->errors() ]);
+            return response()->json(['status' => 500 , 'errors' => "Permission Denied!"]);
         }
     }
 
@@ -85,11 +105,17 @@ class MakeController extends Controller
      */
     public function destroy(string $id)
     {
-        $delete = Make::where('id', $id)->delete();
-        if($delete) {
-            return response()->json(['status' => 200, 'message' => 'Make Deleted Successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('delete-make')) {
+            $delete = Make::where('id', $id)->delete();
+            if($delete) {
+                return response()->json(['status' => 200, 'message' => 'Make Deleted Successfully']);
+            } else {
+                return response()->json(['status' => 500 , 'message' => "something went wrong"]);
+            }
         } else {
-            return response()->json(['status' => 500 , 'message' => "something went wrong"]);
+            return response()->json(['status' => 500 , 'message' => "Permission Denied"]);
         }
     }
 }
