+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
@@ -13,8 +15,15 @@ class UserController extends Controller
      */
     public function index()
     {
-        $user = User::get();
-        return view('admin.manageUser',['user' => $user]);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        $user = $error = null;
+        if($role->hasPermissionTo('view-user')) {
+            $user = User::get();
+        } else {
+            $error = "You don't have permission to access this page";
+        }
+        return view('admin.manageUser',['user' => $user , 'error' => $error]);
     }
 
     /**
@@ -54,17 +63,23 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $validation = Validator::make($request->all(),[
-            'name' => 'required',
-            'email' => 'required|email'
-        ]);
-        if($validation->passes()) {
-            $update = User::where('id',$id)->update(['name'=>$request['name'], 'email' => $request['email'], 'status' => $request['status']]);
-            if($update) {
-                return response()->json(['status' => 200]);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('update-user')) {
+            $validation = Validator::make($request->all(),[
+                'name' => 'required',
+                'email' => 'required|email'
+            ]);
+            if($validation->passes()) {
+                $update = User::where('id',$id)->update(['name'=>$request['name'], 'email' => $request['email'], 'status' => $request['status']]);
+                if($update) {
+                    return response()->json(['status' => 200]);
+                }
+            } else {
+                return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
             }
         } else {
-            return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
+            return response()->json(['status' => 500 , 'errors' => "You don'h have access to update user"]);
         }
     }
 
@@ -73,11 +88,17 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        $delete = User::where('id',$id)->delete();
-        if($delete){
-            return response()->json(['status' => 200 , 'message' => 'User Deleted Successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('delete-user')) {
+            $delete = User::where('id',$id)->delete();
+            if($delete){
+                return response()->json(['status' => 200 , 'message' => 'User Deleted Successfully']);
+            } else {
+                return response()->json(['status' => 500 , 'message' => 'Failed to Delete']);
+            }
         } else {
-            return response()->json(['status' => 500 , 'message' => 'Failed to Delete']);
+            return response()->json(['status' => 500 , 'message' => "You don't have permission to delete user"]);
         }
     }
 }
