+++ app/Http/Controllers/YearController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Year;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class YearController extends Controller
@@ -13,8 +15,15 @@ class YearController extends Controller
      */
     public function index()
     {
-        $year = Year::all();
-        return view('admin.manageYear',['year' => $year]);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        $year = $error = null;
+        if($role->hasPermissionTo('view-year')){
+            $year = Year::all();
+        } else {
+            $error = "Permission Denied";
+        }
+        return view('admin.manageYear',['year' => $year , 'error' => $error]);
     }
 
     /**
@@ -30,16 +39,22 @@ class YearController extends Controller
      */
     public function store(Request $request)
     {
-        $validation = Validator::make($request->all(),[
-            'year' => 'required|unique:year,year|integer',
-        ]);
-        if($validation->passes()) {
-            $insert = Year::create(['year' => $request->year]);
-            if($insert) {
-                return response()->json(['status' => 200 , 'message' => 'Year added successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('create-year')) {
+            $validation = Validator::make($request->all(),[
+                'year' => 'required|unique:year,year|integer',
+            ]);
+            if($validation->passes()) {
+                $insert = Year::create(['year' => $request->year]);
+                if($insert) {
+                    return response()->json(['status' => 200 , 'message' => 'Year added successfully']);
+                }
+            } else {
+                return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
             }
         } else {
-            return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
+               return response()->json(['status' => 500, 'errors' => "Permission Denied"]);
         }
     }
 
@@ -64,16 +79,22 @@ class YearController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $validation = Validator::make($request->all(),[
-            'year' => 'required|unique:year,year,'.$id,
-        ]);
-        if($validation->passes()) {
-            $update = Year::where('id',$id)->update(['year' => $request->year]);
-            if($update) {
-                return response()->json(['status' => 200 , 'message' => 'Year updated successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('update-year')) {
+            $validation = Validator::make($request->all(),[
+                'year' => 'required|unique:year,year,'.$id,
+            ]);
+            if($validation->passes()) {
+                $update = Year::where('id',$id)->update(['year' => $request->year]);
+                if($update) {
+                    return response()->json(['status' => 200 , 'message' => 'Year updated successfully']);
+                }
+            } else {
+                return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
             }
         } else {
-            return response()->json(['status' => 500 , 'errors' => $validation->errors()]);
+            return response()->json(['status' => 400 , 'errors' => "Permission Denied"]);
         }
     }
 
@@ -82,11 +103,17 @@ class YearController extends Controller
      */
     public function destroy(string $id)
     {
-        $delete = Year::where('id', $id)->delete();
-        if($delete) {
-            return response()->json(['status' => 200, 'message' => 'Year Deleted Successfully']);
+        $admin = Auth::guard('admin')->user();
+        $role = Role::find($admin->id);
+        if($role->hasPermissionTo('delete-year')){
+            $delete = Year::where('id', $id)->delete();
+            if($delete) {
+                return response()->json(['status' => 200, 'message' => 'Year Deleted Successfully']);
+            } else {
+                return response()->json(['status' => 500 , 'message' => "Failed To Delete"]);
+            }
         } else {
-            return response()->json(['status' => 500 , 'message' => "Failed To Delete"]);
+            return response()->json(['status' => 500 , 'message' => "Permission Denied"]);
         }
     }
 }
